Fine-Tuning/requirements.txt
ftfy
regex
tqdm
git+https://github.com/openai/CLIP.git
matplotlib
pandas
numpy
pytorch-ignite
tensorboardX
scipy
timm